refactor: separate menu states
